requirements.txt
gradio>=5.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
faiss-cpu>=1.8.0
pypdf>=5.0.0
python-dotenv>=1.0.1
